+++ cpp/3L-VehicleRouting/VehicleRouting/include/VehicleRouting/IntraImprovement/DeleteEmptyRoutes.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "LocalSearchOperatorBase.h"
+#include "Improvement/LocalSearchOperatorBase.h"
 
 namespace VehicleRouting
 {
