+++ cpp/3L-VehicleRouting/VehicleRouting/src/IntraImprovement/TwoOpt.cpp
@@ -1,5 +1,4 @@
-#include "Improvement/TwoOpt.h"
-#include <algorithm>
+#include "IntraImprovement/TwoOpt.h"
 
 namespace VehicleRouting
 {
